requirements.txt
streamlit
transformers>=4.35.0
torch==2.1.0
googletrans==4.0.0-rc1
newspaper3k
sentencepiece
lxml[html_clean]
lxml_html_clean
